GET method actions, better functional separation and unit tests
Improve separation of the functionality, create new interfaces that are
simpler and easy to mock. This allows better testing of the various
components.

- LedgerExecutor handles the execution on the ledger (send transaction,
do a payment), is aware of wallets, blockchain api. It also handles
fees and gas cost
- Datastore only returns data, is aware of wallets
- ActionExecutor is not aware of wallets/blockchain concepts, it only
selects the action and delegate to LedgerExecutor
- Now, actionExecutor.spec only tests the actionExecutor, the datastore
and LedgerExecutor are mocked and tested separately

full Workflow tests everything together but mock the blockchain node.
The wallet tests ensure that the interaction with the blockchain is
correct.